simplify getter/setter templates & refactor test suite
