+++ include/werkzeugkiste/config/types.h
@@ -314,17 +314,19 @@ constexpr const char *TypeName() {
 WZKREG_TNSPEC(bool)
 WZKREG_TNSPEC(char)
 WZKREG_TNSPEC(unsigned char)
-WZKREG_TNSPEC(short)
-WZKREG_TNSPEC(unsigned short)
-WZKREG_TNSPEC(int)
-WZKREG_TNSPEC(unsigned int)
-WZKREG_TNSPEC(long int)
-WZKREG_TNSPEC(unsigned long int)
+WZKREG_TNSPEC(int16_t)
+WZKREG_TNSPEC(uint16_t)
+WZKREG_TNSPEC(int32_t)
+WZKREG_TNSPEC(uint32_t)
+WZKREG_TNSPEC(int64_t)
+WZKREG_TNSPEC(uint64_t)
 WZKREG_TNSPEC(float)
 WZKREG_TNSPEC(double)
 
 WZKREG_TNSPEC(date)
 WZKREG_TNSPEC(time)
+WZKREG_TNSPEC(time_offset)
+WZKREG_TNSPEC(date_time)
 
 WZKREG_TNSPEC_STR(std::string, string)
 WZKREG_TNSPEC_STR(std::string_view, string_view)
